Cast variable to int.

+++ src/HyperMediaPlayer.java
@@ -82,7 +82,7 @@ public class HyperMediaPlayer {
 		    			videoFrame.repaint();  
 		    			
 		    			try {
-		    				videoStack.getFirst().setCurrentFrame(Math.floor(videoStack.getFirst().getClip().getMicrosecondPosition()
+		    				videoStack.getFirst().setCurrentFrame((int)Math.floor(videoStack.getFirst().getClip().getMicrosecondPosition()
 				    				/((double)videoStack.getFirst().getClip().getMicrosecondLength()/(double)videoStack.getFirst().getDuration()));
 		    			}catch(Exception ex) {}
 		    			frameCounter++;
